backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.64",
    "fastapi>=0.120.4",
    "mangum>=0.19.0",
    "openai>=2.6.1",
    "pydantic[email]>=2.12.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
